refactor: move authOptions to a separate config file for better organization and maintainability
