+++ nextjs-temp/src/app/api/analyze-complete/route.ts
@@ -1,7 +1,7 @@
 import { NextRequest, NextResponse } from 'next/server';
 import { getServerSession } from 'next-auth';
 import { PrismaClient } from '@prisma/client';
-import { authOptions } from '@/app/api/auth/[...nextauth]/route';
+import { authOptions } from '@/config/auth';
 import { GeminiService } from '@/services/gemini';
 
 const prisma = new PrismaClient();
